+++ R/MeasureClustSimple.R
@@ -29,6 +29,9 @@ MeasureClustSimple = R6Class(
     .requires_k2 = NULL,
 
     .score = function(prediction, task, ...) {
+      if (length(prediction$row_ids) == 0L) {
+        return(NaN)
+      }
       if (private$.requires_k2 && length(unique(prediction$partition)) < 2L) {
         return(NaN)
       }